+++ server/php/Genghis/Api.php
@@ -207,10 +207,11 @@ class Genghis_Api extends Genghis_App
     {
         switch ($method) {
             case 'GET':
-                $query = (string) $this->getQueryParam('q', '');
-                $page  = (int) $this->getQueryParam('page', 1);
+                $query  = (string) $this->getQueryParam('q',      null);
+                $fields = (string) $this->getQueryParam('fields', null);
+                $page   = (int)    $this->getQueryParam('page',   1);
 
-                return $this->servers[$server][$db][$coll]->findDocuments($query, $page);
+                return $this->servers[$server][$db][$coll]->findDocuments($query, $fields, $page);
 
             case 'POST':
                 return $this->servers[$server][$db][$coll]->insert($this->getRequestData());